+++ CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to
 
 ### Fixed
 
+- Fix ESC key closing IDE when pressed on adaptor modal
+  [#3978](https://github.com/OpenFn/lightning/issues/3978)
 - Fix run/save-and-run keystroke mapping for canvas & IDE
   [#3902](https://github.com/OpenFn/lightning/issues/3902) &
   [#3903](https://github.com/OpenFn/lightning/issues/3903)